Refactored copilot-usage into modular model-usage plugin
