+++ .config/opencode/plugins/model-usage.config.json
