+++ .config/opencode/plugins/model-usage/sidebar.tsx
@@ -1,67 +1,30 @@
 /** @jsxImportSource @opentui/solid */
-import type { TuiPlugin, TuiPluginModule, TuiPluginApi } from "@opencode-ai/plugin/tui"
+import type { TuiPluginApi } from "@opencode-ai/plugin/tui"
 import type { EventMessageUpdated } from "@opencode-ai/sdk/v2"
 import { createSignal, createEffect, createMemo, onCleanup } from "solid-js"
-import { mkdirSync, appendFileSync, existsSync, readFileSync } from "node:fs"
-import { homedir } from "node:os"
+import type { CopilotConfig, CopilotQuotaInfo, GoQuotaInfo, MessagePart } from "./types"
+import {
+  log as logFn,
+  DEBUG,
+  isSupportedModel, calculateMessageMultiplier,
+  roundUsage, buildProgressBar, getUsageColor, getQuotaLabel,
+  formatDuration, isModelDeprecated,
+  splitCost,
+  logPath,
+} from "./helpers"
+import { appendFileSync } from "node:fs"
+
+const log = logFn
+import { fetchQuotaInfo, fetchGoQuota } from "./quota"
 
 const QUOTA_REFRESH_MS = 5 * 60 * 1000
 const MAX_POLL_ATTEMPTS = 30
 const PLUGIN_VERSION = "v35"
 
-interface CopilotConfig {
-  modelMultipliers: Record<string, number>
-  deprecated: string[]
-}
-
-interface CopilotQuotaInfo {
-  percentRemaining: number
-  entitlement: number
-  remaining: number
-  overageCount: number
-  overagePermitted: boolean
-  unlimited: boolean
-  planType: "free" | "paid"
-  quotaType: "premium" | "ai_credits"
-  tokenBasedBilling: boolean
-}
-
-interface GoQuotaBar {
-  usagePercent: number
-  resetInSec: number
-}
-
-interface GoQuotaInfo {
-  rolling: GoQuotaBar | null
-  weekly: GoQuotaBar | null
-  monthly: GoQuotaBar | null
-}
-
-interface MessagePart {
-  type: string
-  synthetic?: boolean
-}
-
-const DEBUG = process.env.OPENCODE_COPILOT_DEBUG === "true"
-const logsDir = new URL("./logs", import.meta.url).pathname
-const logPath = new URL(`./logs/log_copilot_plugin_${Date.now()}.log`, import.meta.url).pathname
-if (DEBUG) mkdirSync(logsDir, { recursive: true })
-
-const log = (...args: unknown[]) => {
-  if (!DEBUG) return
-  try {
-    const line = `[${new Date().toISOString()}] ${args.map(String).join(" ")}\n`
-    appendFileSync(logPath, line)
-  } catch {
-    // ignore
-  }
-}
-
-log(`=== copilot-usage ${PLUGIN_VERSION} loaded ===`)
-log("logPath:", logPath)
+log(`=== usage-sidebar ${PLUGIN_VERSION} loaded ===`)
 
 async function loadConfig(): Promise<CopilotConfig> {
-  const configPath = new URL("./copilot-usage.config.json", import.meta.url).pathname
+  const configPath = new URL("../model-usage.config.json", import.meta.url).pathname
   log("loadConfig: trying path:", configPath)
   try {
     const configFile = Bun.file(configPath)
@@ -86,82 +49,6 @@ async function loadConfig(): Promise<CopilotConfig> {
     modelMultipliers: {},
     deprecated: [],
   }
-}
-
-function getModelId(modelName: string): string | null {
-  if (!modelName) return null
-  const parts = modelName.split("/")
-  if (parts.length < 2) return null
-  const provider = parts[0]
-  if (!provider.includes("copilot") && provider !== "opencode-go") return null
-  return parts.slice(1).join("/")
-}
-
-function getMultiplier(modelName: string, config: CopilotConfig): number {
-  const modelId = getModelId(modelName)
-  if (!modelId) return 1.0
-  if (config.modelMultipliers[modelId] !== undefined) {
-    return config.modelMultipliers[modelId]
-  }
-  const normalized = modelId.toLowerCase()
-  for (const [key, value] of Object.entries(config.modelMultipliers)) {
-    if (key.toLowerCase() === normalized) return value
-  }
-  return 1.0
-}
-
-function isModelDeprecated(modelId: string, config: CopilotConfig): boolean {
-  if (!modelId) return false
-  const normalized = modelId.toLowerCase()
-  return config.deprecated.some((d) => d.toLowerCase() === normalized)
-}
-
-function isSupportedModel(modelName: string): boolean {
-  if (!modelName) return false
-  const lower = modelName.toLowerCase()
-  return lower.includes("copilot") || lower.includes("opencode-go")
-}
-
-function isSyntheticMessage(parts: MessagePart[]): boolean {
-  // Messages with a "compaction" type part are the real compaction request — always count them.
-  if (parts.some((p) => p.type === "compaction")) return false
-  // The synthetic *continuation* message (auto-created after compaction to resume the chat)
-  // has ONLY {type:"text", synthetic:true} parts and nothing else.
-  // Real user messages may also contain synthetic text parts (e.g. the first message in a
-  // session after context refill), but they always have at least one non-synthetic part too
-  // (e.g. a plain "text" part, a "file" part, etc.).
-  const syntheticTextParts = parts.filter((p) => p.type === "text" && p.synthetic === true)
-  if (syntheticTextParts.length === 0) return false
-  const nonSyntheticParts = parts.filter((p) => !(p.type === "text" && p.synthetic === true))
-  return nonSyntheticParts.length === 0
-}
-
-function calculateMessageMultiplier(
-  msgId: string,
-  parts: MessagePart[],
-  model: string | null,
-  isFreePlan: boolean,
-  config: CopilotConfig,
-  messageMultipliers: Map<string, number>
-): number {
-  if (isSyntheticMessage(parts)) {
-    // Don't store in map — storing 0 would permanently blacklist the message ID,
-    // preventing any future retry if parts change (e.g. message.updated fires again).
-    log("calculateMessageMultiplier: skipping synthetic message", msgId)
-    return 0
-  }
-
-  if (!model || getModelId(model) === null) {
-    // Don't store in map — allows retry when the model becomes available on the next event fire.
-    log("calculateMessageMultiplier: no copilot model for", msgId, "model:", model)
-    return 0
-  }
-
-  const multiplier = isFreePlan ? 1.0 : getMultiplier(model, config)
-  log("calculateMessageMultiplier:", msgId, "model:", model, "multiplier:", multiplier)
-  // Only store positive results — the deduplication guard in message.updated checks this map.
-  messageMultipliers.set(msgId, multiplier)
-  return multiplier
 }
 
 function getActiveModel(api: TuiPluginApi, sessionID: string): string | null {
@@ -212,245 +99,8 @@ function getActiveModel(api: TuiPluginApi, sessionID: string): string | null {
   return null
 }
 
-async function fetchQuotaInfo(token: string): Promise<CopilotQuotaInfo | null> {
-  try {
-    log("fetchQuotaInfo: fetching quota from github API")
-    const response = await fetch("https://api.github.com/copilot_internal/user", {
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: "application/json",
-        "X-GitHub-Api-Version": "2026-06-01",
-        "User-Agent": "opencode-copilot-usage-plugin",
-      },
-    })
-
-    log("fetchQuotaInfo: response status:", response.status)
-    if (!response.ok) {
-      log("fetchQuotaInfo: response not ok, body:", await response.text().catch(() => "failed to read"))
-      return null
-    }
-
-    const data = await response.json()
-    log("fetchQuotaInfo: copilot_plan:", data?.copilot_plan)
-
-    // Paid plan: try premium_models first, fall back to premium_interactions.
-    // AI Credits detection: check snapshot.token_based_billing or root token_based_billing flag
-    // (GitHub migrated from count-based "premium requests" to token-based "AI Credits").
-    const snapshot = data?.quota_snapshots?.premium_models ?? data?.quota_snapshots?.premium_interactions
-    if (snapshot) {
-      return {
-        percentRemaining: snapshot.percent_remaining ?? 100,
-        entitlement: snapshot.entitlement ?? 0,
-        remaining: snapshot.remaining ?? 0,
-        overageCount: snapshot.overage_count ?? 0,
-        overagePermitted: snapshot.overage_permitted ?? false,
-        unlimited: snapshot.unlimited ?? false,
-        planType: "paid",
-        quotaType: (data?.quota_snapshots?.premium_models || snapshot.token_based_billing || data?.token_based_billing) ? "ai_credits" : "premium",
-        tokenBasedBilling: !!(snapshot.token_based_billing || data?.token_based_billing),
-      }
-    }
-
-    // Free plan: limited_user_quotas + monthly_quotas
-    const chatRemaining = data?.limited_user_quotas?.chat
-    const chatMonthly = data?.monthly_quotas?.chat
-    if (chatRemaining != null && chatMonthly != null && chatMonthly > 0) {
-      const percentRemaining = Math.round((chatRemaining / chatMonthly) * 1000) / 10
-      log("fetchQuotaInfo: free plan, chat remaining:", chatRemaining, "/", chatMonthly, "=", percentRemaining.toFixed(1) + "%")
-      return {
-        percentRemaining,
-        entitlement: chatMonthly,
-        remaining: chatRemaining,
-        overageCount: 0,
-        overagePermitted: false,
-        unlimited: false,
-        planType: "free",
-        quotaType: "premium",
-        tokenBasedBilling: false,
-      }
-    }
-
-    log("fetchQuotaInfo: no quota data found")
-    return null
-  } catch (err) {
-    log("fetchQuotaInfo: error:", String(err))
-    return null
-  }
-}
-
-function formatDuration(seconds: number): string {
-  if (seconds < 60) return `${seconds}s`
-  if (seconds < 3600) return `${Math.floor(seconds / 60)}m`
-  if (seconds < 86400) {
-    const h = Math.floor(seconds / 3600)
-    const m = Math.floor((seconds % 3600) / 60)
-    return m > 0 ? `${h}h ${m}m` : `${h}h`
-  }
-  const d = Math.floor(seconds / 86400)
-  const h = Math.floor((seconds % 86400) / 3600)
-  return h > 0 ? `${d}d ${h}h` : `${d}d`
-}
-
-function getGoAuth(): { type: "bearer" | "cookie"; value: string } | null {
-  // Cookie auth first — required for fetching the web console page
-  const cookie = process.env.OPENCODE_GO_AUTH_COOKIE
-  if (cookie) return { type: "cookie", value: cookie }
-
-  // Bearer token fallback (inference key, may not work for web page)
-  try {
-    const authPath = `${homedir()}/.local/share/opencode/auth.json`
-    if (existsSync(authPath)) {
-      const raw = readFileSync(authPath, "utf8")
-      const auth = JSON.parse(raw)
-      const key = auth?.["opencode-go"]?.key
-      if (key && auth["opencode-go"]?.type === "api") {
-        return { type: "bearer", value: key }
-      }
-    }
-  } catch {
-    // ignore
-  }
-
-  const apiKey = process.env.OPENCODE_GO_API_KEY
-  if (apiKey) return { type: "bearer", value: apiKey }
-
-  return null
-}
-
-async function fetchGoQuota(): Promise<GoQuotaInfo | null> {
-  const workspaceId = process.env.OPENCODE_GO_WORKSPACE_ID
-  if (!workspaceId) return null
-
-  const auth = getGoAuth()
-  if (!auth) return null
-
-  const patterns = {
-    rolling: /rollingUsage:\$R\[\d+\]=(\{[^}]+\})/,
-    weekly: /weeklyUsage:\$R\[\d+\]=(\{[^}]+\})/,
-    monthly: /monthlyUsage:\$R\[\d+\]=(\{[^}]+\})/,
-  }
-
-  try {
-    const headers: Record<string, string> = {
-      "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36",
-      Accept: "text/html,application/xhtml+xml",
-    }
-    if (auth.type === "bearer") {
-      headers["Authorization"] = `Bearer ${auth.value}`
-    } else {
-      headers["Cookie"] = `auth=${auth.value}`
-    }
-
-    const response = await fetch(
-      `https://opencode.ai/workspace/${encodeURIComponent(workspaceId)}/go`,
-      { headers },
-    )
-
-    if (!response.ok) return null
-
-    const html = await response.text()
-    const usage: GoQuotaInfo = { rolling: null, weekly: null, monthly: null }
-
-    for (const [key, pattern] of Object.entries(patterns)) {
-      const match = html.match(pattern)
-      if (match) {
-        try {
-          const jsonStr = match[1].replace(/([{,]\s*)([a-zA-Z_][a-zA-Z0-9_]*)(\s*:)/g, '$1"$2"$3')
-          usage[key as keyof GoQuotaInfo] = JSON.parse(jsonStr)
-        } catch {
-          // ignore individual parse failures
-        }
-      }
-    }
-
-    return usage
-  } catch {
-    return null
-  }
-}
-
-function buildProgressBar(percentage: number, width: number = 20): string {
-  const filled = Math.min(Math.round((percentage / 100) * width), width)
-  const empty = width - filled
-  return "\u2588".repeat(filled) + "\u2591".repeat(empty)
-}
-
-function roundUsage(value: number): number {
-  let rounded = Math.round(value * 100) / 100
-  // Handle imprecise multiplier artifacts (0.33 ≈ 1/3):
-  // 2×0.33=0.65→0.67, 4×0.33=1.32→1.33, etc.
-  let cents = Math.round((rounded % 1) * 100)
-  if (cents > 0 && cents < 100 && cents % 33 === 32) {
-    rounded = Math.round((rounded + 0.01) * 100) / 100
-    cents = Math.round((rounded % 1) * 100)
-  }
-  // Handle .99 → round up (3×0.33=0.99→1.00)
-  if (cents === 99) {
-    return Math.ceil(rounded)
-  }
-  return rounded
-}
-
-function getUsageColor(percentage: number): string {
-  if (percentage > 100) return "#ef4444"
-  if (percentage > 90) return "#ef4444"
-  if (percentage > 75) return "#eab308"
-  return "#22c55e"
-}
-
-function getQuotaLabel(quota: CopilotQuotaInfo): string {
-  if (quota.planType === "free") return "chat requests"
-  if (quota.quotaType === "ai_credits") return "AI Credits"
-  return "premium requests"
-}
-
-function splitCost(
-  deltaInput: number,
-  deltaCacheRead: number,
-  deltaOutput: number,
-  deltaCost: number,
-  modelId: string,
-  provider: ReadonlyArray<any>
-): { inputCost: number; outputCost: number } {
-  let inputPrice = 0
-  let outputPrice = 0
-  let cacheReadPrice = 0
-
-  for (const p of provider) {
-    const model = p.models?.[modelId]
-    if (model?.cost) {
-      inputPrice = model.cost.input ?? 0
-      outputPrice = model.cost.output ?? 0
-      cacheReadPrice = model.cost.cache?.read ?? 0
-      break
-    }
-  }
-
-  // If no pricing found, fall back to raw token proportional split
-  if (inputPrice === 0 && outputPrice === 0) {
-    const totalTok = deltaInput + deltaCacheRead + deltaOutput
-    if (totalTok === 0) return { inputCost: 0, outputCost: 0 }
-    return {
-      inputCost: deltaCost * (deltaInput + deltaCacheRead) / totalTok,
-      outputCost: deltaCost * deltaOutput / totalTok,
-    }
-  }
-
-  // Price-weighted split
-  const inputWeight = (deltaInput * inputPrice + deltaCacheRead * cacheReadPrice) / 1_000_000
-  const outputWeight = (deltaOutput * outputPrice) / 1_000_000
-  const totalWeight = inputWeight + outputWeight
-
-  if (totalWeight === 0) return { inputCost: 0, outputCost: 0 }
-
-  return {
-    inputCost: deltaCost * inputWeight / totalWeight,
-    outputCost: deltaCost * outputWeight / totalWeight,
-  }
-}
-
-function CopilotUsageSidebar(props: { api: TuiPluginApi; session_id: string }) {
-  log("CopilotUsageSidebar: rendering! session_id:", props.session_id)
+function UsageSidebar(props: { api: TuiPluginApi; session_id: string }) {
+  log("UsageSidebar: rendering! session_id:", props.session_id)
   const githubToken = process.env.GITHUB_TOKEN ?? ""
   const hasToken = !!githubToken
 
@@ -964,7 +614,7 @@ function CopilotUsageSidebar(props: { api: TuiPluginApi; session_id: string }) {
   const isCopilotActive = createMemo(() => currentModel()?.toLowerCase().includes("copilot") ?? false)
   const isOpenCodeGoActive = createMemo(() => currentModel()?.toLowerCase().includes("opencode-go") ?? false)
 
-  log("CopilotUsageSidebar: render, isSupported:", isSupported(), "activeModel:", currentModel())
+  log("UsageSidebar: render, isSupported:", isSupported(), "activeModel:", currentModel())
 
   return (
     <box flexDirection="column" gap={0}>
@@ -1061,23 +711,4 @@ function CopilotUsageSidebar(props: { api: TuiPluginApi; session_id: string }) {
   )
 }
 
-const tui: TuiPlugin = async (api) => {
-  log("copilot-usage: tui function called")
-  api.slots.register({
-    order: 175,
-    slots: {
-      sidebar_content(_ctx, props) {
-        log("copilot-usage: rendering sidebar_content, session_id:", props.session_id)
-        return <CopilotUsageSidebar api={api} session_id={props.session_id} />
-      },
-    },
-  })
-  log("copilot-usage: slot registered")
-}
-
-const plugin: TuiPluginModule & { id: string } = {
-  id: "copilot-usage",
-  tui,
-}
-
-export default plugin
+export default UsageSidebar
