+++ .config/opencode/tui.json
@@ -1,4 +1,4 @@
 {
   "$schema": "https://opencode.ai/tui.json",
-  "plugin": ["./plugins/copilot-usage.tsx", "./plugins/usage-command.tsx"]
+  "plugin": ["./plugins/model-usage.tsx"]
 }
